services/etl/requirements.txt
httpx>=0.27.0
pandas>=2.2.0
python-dotenv>=1.0.0
supabase>=2.5.0
resend>=0.8.0
pydantic>=2.7.0
pytz>=2024.1
chardet>=5.2.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
respx>=0.21.0
factory-boy>=3.3.0
freezegun>=1.4.0
